requirements.txt
fastapi>=0.115.9
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
openai==1.78.0
chromadb==1.0.5
pypdf==5.4.0
python-docx==1.1.2
pydantic==2.11.4
